+++ bioit_bigsdb_scripts/components/maininserter.py
@@ -1,7 +1,7 @@
 import dataclasses
 import datetime
-import logging
 import socket
+from logging import getLogger
 from typing import Any, Dict
 
 from bioit_mongodb_scripts.model.json_model import JsonReportDict
@@ -9,6 +9,8 @@ from .json_superclass import JsonSuperClass
 from .psql import TblEavBoolean, TblEavFields, TblEavInt, TblEavText, TblEavTextHidden, TblHistory, TblIsolates
 from ..utils.html_tbl_templates import HtmlAntiviralAssociationsTableBuilder, HtmlAntiviralMutationsTableBuilder, HtmlRefSelectionTableBuilder, HtmlReportBuilder, HtmlSnpLineageTableBuilder
 from ..utils.url_helper import UrlHelper
+
+logger = getLogger(__name__)
 
 
 @dataclasses.dataclass
@@ -108,7 +110,7 @@ class MainInserter(JsonSuperClass):
                 self.isolates_psql_tbl.add_validation((self._json_report_dict['validation']['type'], self._json_report_dict['validation']['curator'],
                                                        datetime.datetime.strptime(self._json_report_dict['validation']['date'], '%d/%m/%Y - %X').strftime('%Y-%m-%d'),
                                                        str(context.isolate_id)))
-            logging.info('Metadata insertion successful')
+            logger.info('Metadata insertion successful')
 
     def __insert_species_specific_metadata(self, context: MainInserterContext) -> None:
         """