+++ lib/BIGSdb/Page.pm
@@ -177,6 +177,7 @@ sub _get_javascript_paths {
 		push @$js, { src => "$relative_js_path/jquery-ui.min.js", defer   => 1, version => '1.12.1' };
 		push @$js, { src => "$relative_js_path/bigsdb.min.js",    defer   => 1, version => '20231205' };
 		push @$js, { src => "$relative_js_path/hera_dynamic_query.js"};
+		push @$js, { src => "$relative_js_path/hera_utils.js"};
 		push @$js, { src => "$relative_js_path/jszip.js"};
 		if ( !$self->{'config'}->{'no_cookie_consent'} && !$self->{'curate'} && $self->{'instance'} ) {
 			push @$js, { src => "$relative_js_path/cookieconsent.min.js", defer => 1 };
@@ -822,7 +823,8 @@ sub _get_stylesheets {
 	my $system  = $self->{'system'};
 	my $version = '20250318';
 	my @filenames;
-	push @filenames, q(dropzone.css)                                          if $self->{'dropzone'};
+	push @filenames, q(hera_style.css);
+    push @filenames, q(dropzone.css)                                          if $self->{'dropzone'};
 	push @filenames, q(billboard.min.css)                                     if $self->{'billboard'};
 	push @filenames, q(pivot.min.css)                                         if $self->{'pivot'};
 	push @filenames, qw(jquery.multiselect.css jquery.multiselect.filter.css) if $self->{'jQuery.multiselect'};